Old sega from geometry and segang file, gretina stuff to sort geant simulations

+++ libraries/GROOT/GH1.cxx
@@ -48,13 +48,13 @@ double GH1::FitEval(double *dim,double *par) {
   //printf("par = 0x%08x\n",par); fflush(stdout);
   double scale = par[0];
 
-  int binNum = GetXaxis()->FindBin(x); //gHist->GetBin() does not respect rebinning.
+  double binNum = GetXaxis()->FindBin(x); //gHist->GetBin() does not respect rebinning.
 
-  int nBins = GetNbinsX();
-  int kevPerBin = GetXaxis()->GetXmax()/nBins;
-  int curBinX = GetBinCenter(binNum);
-  int nextBinX = GetBinCenter(binNum+1);
-  int prevBinX = GetBinCenter(binNum-1);
+  double nBins = GetNbinsX();
+  double kevPerBin = GetXaxis()->GetXmax()/nBins;
+  double curBinX = GetBinCenter(binNum);
+  double nextBinX = GetBinCenter(binNum+1);
+  double prevBinX = GetBinCenter(binNum-1);
 
   if (x > prevBinX && x <= curBinX){
     double leftDiff = x - prevBinX;
@@ -87,13 +87,13 @@ double GH1::FitEval2(double *dim,double *par) {
   //printf("par = 0x%08x\n",par); fflush(stdout);
   double scale = par[0];
 
-  int binNum = GetXaxis()->FindBin(x); //gHist->GetBin() does not respect rebinning.
+  double binNum = GetXaxis()->FindBin(x); //gHist->GetBin() does not respect rebinning.
 
-  int nBins = GetNbinsX();
-  int kevPerBin = GetXaxis()->GetXmax()/nBins;
-  int curBinX = GetBinCenter(binNum);
-  int nextBinX = GetBinCenter(binNum+1);
-  int prevBinX = GetBinCenter(binNum-1);
+  double nBins = GetNbinsX();
+  double kevPerBin = GetXaxis()->GetXmax()/nBins;
+  double curBinX = GetBinCenter(binNum);
+  double nextBinX = GetBinCenter(binNum+1);
+  double prevBinX = GetBinCenter(binNum-1);
 
   if (x > prevBinX && x <= curBinX){
     double leftDiff = x - prevBinX;
